# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Приёмка API",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "maxInitData": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Max-Init-Data",
        "description": "Исходная строка window.WebApp.initData"
      }
    },
...
  },
  "paths": {
...
    "/api/me": {
      "get": {
        "summary": "Текущий пользователь MAX",
        "tags": [
          "Identity"
        ],
        "security": [
          {
            "maxInitData": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "user",
                    "auth_date",
                    "houses"
                  ],
                  "properties": {
                    "query_id": {
                      "type": "string"
                    },
                    "ip": {
                      "type": "string"
                    },
                    "auth_date": {
                      "type": "integer"
                    },
                    "user": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "id",
                        "maxUserId",
                        "isAdmin",
                        "first_name",
                        "last_name",
                        "username",
                        "language_code",
                        "photo_url"
                      ],
                      "properties": {
                        "id": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "maxUserId": {
                          "type": "string",
                          "pattern": "^[1-9][0-9]*$"
                        },
                        "isAdmin": {
                          "type": "boolean"
                        },
                        "first_name": {
                          "type": "string"
                        },
                        "last_name": {
                          "type": "string"
                        },
                        "username": {
                          "type": "string",
                          "nullable": true
                        },
                        "language_code": {
                          "type": "string"
                        },
                        "photo_url": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "chat": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "id",
                        "type"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "pattern": "^(?:0|-?[1-9][0-9]*)$"
                        },
                        "type": {
                          "type": "string",
                          "enum": [
                            "DIALOG",
                            "CHAT",
                            "CHANNEL"
                          ]
                        }
                      }
                    },
                    "start_param": {
                      "type": "string"
                    },
                    "houses": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": [
                          "id",
                          "address",
                          "role",
                          "status",
                          "permissions"
                        ],
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "address": {
                            "type": "string"
                          },
                          "role": {
                            "type": "string",
                            "enum": [
                              "RESIDENT",
                              "COUNCIL_MEMBER",
                              "CHAIRMAN",
                              "EXECUTOR"
                            ]
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "PENDING",
                              "ACTIVE",
                              "REJECTED"
                            ]
                          },
                          "permissions": {
                            "type": "object",
                            "additionalProperties": false,
                            "required": [
                              "viewWorks",
                              "viewObservations",
                              "viewHouseChat",
                              "createObservation",
                              "commentWork",
                              "watchWork",
                              "manageHouseChat",
                              "assignInspectors",
                              "performInspection",
                              "reportRemediation",
                              "confirmWorkResult",
                              "reviewJoinRequests"
                            ],
                            "properties": {
                              "viewWorks": {
                                "type": "boolean"
                              },
                              "viewObservations": {
                                "type": "boolean"
                              },
                              "viewHouseChat": {
                                "type": "boolean"
                              },
                              "createObservation": {
                                "type": "boolean"
                              },
                              "commentWork": {
                                "type": "boolean"
                              },
                              "watchWork": {
                                "type": "boolean"
                              },
                              "manageHouseChat": {
                                "type": "boolean"
                              },
                              "assignInspectors": {
                                "type": "boolean"
                              },
                              "performInspection": {
                                "type": "boolean"
                              },
                              "reportRemediation": {
                                "type": "boolean"
                              },
                              "confirmWorkResult": {
                                "type": "boolean"
                              },
                              "reviewJoinRequests": {
                                "type": "boolean"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/media": {
      "post": {
        "tags": [
          "Media"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "maxInitData": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "id"
                  ],
                  "properties": {
                    "id": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/photo/{key}": {
      "get": {
        "tags": [
          "Media"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 32,
              "maximum": 2048
            },
            "in": "query",
            "name": "w",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 32,
              "maximum": 2048
            },
            "in": "query",
            "name": "h",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "cover",
                "contain"
              ]
            },
            "in": "query",
            "name": "fit",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9]{12,16}$"
            },
            "in": "path",
            "name": "key",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "image/png": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "image/webp": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "image/png": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "image/webp": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/doc/{key}.pdf": {
      "get": {
        "tags": [
          "Documents"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^(?:[a-z0-9]{20}|[a-f0-9]{40})$"
            },
            "in": "path",
            "name": "key",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}